Change health entry structure
