+++ App.js
@@ -24,7 +24,7 @@ import ViewJournalEntry from './app/journal/screens/ViewJournalEntry';
 import WordCloudDisplay from './app/journal/screens/WordCloudDisplay';
 import { Goals } from './app/goals/screens/Goals';
 import { GoalComplete } from './app/goals/screens/GoalComplete';
-import { HealthEntry1 } from './app/healthentry/screens/HealthEntry';
+import HealthEntry from './app/healthentry/screens/HealthEntry';
 import { ToDoList } from './app/todolist/screens/ToDoList';
 import {
   HistoryHealthEntries,
@@ -168,7 +168,7 @@ function App() {
         />
         <Stack.Screen
           name='HealthEntry1'
-          component={HealthEntry1}
+          component={HealthEntry}
           options={{ header: () => null }}
         />
         <Stack.Screen
