+++ app/healthentry/screens/EntryCompletion.js
@@ -1,0 +1,51 @@
+import React from 'react';
+import {
+  Button,
+  Modal,
+  SafeAreaView,
+  StyleSheet,
+  Text,
+  View,
+  Image,
+} from 'react-native';
+import { TextInput } from 'react-native-gesture-handler';
+import NavBar from '../../shared/components/NavBar';
+
+const EntryCompletion = () => {
+  return (
+    <SafeAreaView>
+      <View>
+        <Image source={require('../../shared/assets/icon.png')} />
+      </View>
+      <View>
+        <Image source={require('../../shared/assets/icon.png')} />
+        <Text>You have watered (plant name) today!</Text>
+        <Text>+5 growth points</Text>
+      </View>
+      <View>
+        <Button title='View Entry' />
+        <Button title='Return to Home' />
+      </View>
+    </SafeAreaView>
+  );
+};
+
+export default EntryCompletion;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F6EFED',
+  },
+  avatar: {
+    width: 75,
+    height: 75,
+    marginRight: 24,
+  },
+  buttons: {
+    marginTop: 10,
+    marginBottom: 10,
+    width: 100,
+    height: 50,
+  },
+});
