+++ app/healthentry/screens/ReviewEntry.js
@@ -1,0 +1,40 @@
+import React from 'react';
+import {
+  Button,
+  Modal,
+  SafeAreaView,
+  StyleSheet,
+  Text,
+  View,
+  Image,
+} from 'react-native';
+import { TextInput } from 'react-native-gesture-handler';
+import NavBar from '../../shared/components/NavBar';
+
+const ReviewEntry = () => {
+  return (
+    <View>
+      <Text></Text>
+    </View>
+  );
+};
+
+export default ReviewEntry;
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F6EFED',
+  },
+  avatar: {
+    width: 75,
+    height: 75,
+    marginRight: 24,
+  },
+  buttons: {
+    marginTop: 10,
+    marginBottom: 10,
+    width: 100,
+    height: 50,
+  },
+});
